feat(carousel): move subcomponents to subfolders #358
