+++ src/tedi/components/content/carousel/index.ts
@@ -1,7 +1,7 @@
 export * from './carousel';
-export * from './carousel-content';
-export * from './carousel-footer';
-export * from './carousel-header';
-export * from './carousel-indicators';
-export * from './carousel-navigation';
+export * from './components/carousel-content/carousel-content';
+export * from './components/carousel-footer/carousel-footer';
+export * from './components/carousel-header/carousel-header';
+export * from './components/carousel-indicators/carousel-indicators';
+export * from './components/carousel-navigation/carousel-navigation';
 export * from './carousel-utils';
